_strnicmp infinite recursion with some versions of string.h
Some versions of string.h define strncasecmp to _strnicmp. Causing the existing implementation to resolve to an infinitely recursive function.

The define from string.h can be copied here to ensure that _strnicmp is used in the event that the version of string.h included does not have the macro.